+++ packages/heypi/src/store/index.ts
@@ -12,6 +12,8 @@ export type {
 	Locks,
 	Message,
 	Messages,
+	ProviderMessage,
+	ProviderMessages,
 	Store,
 	StoredMessage,
 	Thread,